GSE/requirements.txt
pyserial==3.4
pyzmq==17.0.0
PyQt5==5.11.2
qtconsole==4.3.1
jupyter==1.0.0